requirements.txt
google-genai>=1.0.0
pandas>=2.0.0
Pillow>=10.0.0
pydantic>=2.0.0
openpyxl>=3.1.0
